move config to agent-dcs dir
