+++ agent-docs/config.ts
@@ -1,0 +1,2 @@
+export const VECTOR_STORE_NAME = process.env.VECTOR_STORE_NAME || 'docs';
+export const vectorSearchNumber = 20;
