+++ agent-docs/src/agents/doc-processing/docs-orchestrator.ts
@@ -1,6 +1,6 @@
 import type { AgentContext } from '@agentuity/sdk';
 import { processDoc } from './docs-processor';
-import { VECTOR_STORE_NAME } from '../../../../config';
+import { VECTOR_STORE_NAME } from '../../../config';
 import type { SyncPayload, SyncStats } from './types';
 
 /**
